Separate gizmos from pbr viewer, modify first frame of taa effect, remove unnecessary files
